Expenses

See merge request cse1105/2023-2024/teams/oopp-team-08!14

+++ server/src/main/java/server/Event.java
@@ -79,8 +79,8 @@ public class Event {
      * Method to calculate the total expenses of the Event
      * @return total sum of Expenses
      */
-    public int getTotal() {
-        int totalExpenses = 0;
+    public double getTotal() {
+        double totalExpenses = 0;
         for(int i = 0; i < expenses.size(); i++) {
             totalExpenses = totalExpenses + expenses.get(i).getAmount();
         }
